# Improve performance with better CommandInfo cache locking making Invoke-ScriptAnalyzer nearly twice as fast