@rjmholt 2019-03-04
The principle at work here seems not unlike that I wrote for the profile cache (which had to optimise for concurrent callers wanting optimum parallel execution):

https://github.com/PowerShell/PSScriptAnalyzer/blob/dff952255db1bc188d4319a2479e9aed9ab7abc1/PSCompatibilityAnalyzer/Microsoft.PowerShell.CrossCompatibility/Utility/CompatibilityProfileLoader.cs#L108-L130

I think we could easily avoid the duplicated work in (2) with lazy task entries -- it would probably also decrease the amount of logic you've had to write here.

If you'd prefer, I could also give this a shot myself.