@bergmeister 2019-03-04
Hmm, I think our 2 scenarios are still a bit different. I for example have 2 concurrent dictionaries, one is for the results and one is a way of managing and retrieving currently running requests. This is to keep the memory usage minimal and release allocations of finished tasks because my dictionary will potentially hold thousands of objects. I don't see how Lazy would be useful in my case. Also, I am not async yet (this would be another PR using runspaces).
Feel free to give it a go but to me this looks like quite a bit of effort to unify the code. Or maybe I am not understanding correctly how you'd apply your approach for this scenario. I am happy to be convinced otherwise.