@rjmholt 2019-03-05
I had a look and talked about it a bit with @daxian-dbw, who made the point that if we want the result immediately, there may not be a good reason to use a `Task`.

I just wrote a simple implementation around more granular locks in https://github.com/PowerShell/PSScriptAnalyzer/pull/1166, which seems to get the same performance (compared both modules with current `development` and got about a 17% speed up in both your and my implementations).