@rjmholt 2019-03-05
I also did some experimentation with async, since the `PowerShell` type has `BeginInvoke()` and `EndInvoke()` methods, but didn't see much reason to keep them in the code since we can't quite use them yet